+++ radixdlt/src/integration/java/com/radixdlt/consensus/deterministic/bft/synchronous/RandomChannelOrderResponsiveTest.java
@@ -15,7 +15,7 @@
  * language governing permissions and limitations under the License.
  */
 
-package com.radixdlt.consensus.deterministic.synchronous;
+package com.radixdlt.consensus.deterministic.bft.synchronous;
 
 import com.radixdlt.consensus.deterministic.BFTDeterministicTest;
 import java.util.Random;
