+++ radixdlt/src/integration/java/com/radixdlt/consensus/simulation/bft/synchronous/UniformLatencyTest.java
@@ -15,7 +15,7 @@
  * language governing permissions and limitations under the License.
  */
 
-package com.radixdlt.consensus.simulation.synchronous;
+package com.radixdlt.consensus.simulation.bft.synchronous;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
